Fixed the circular import problem
