+++ pennylane_calculquebec/device_exception.py
@@ -1,0 +1,5 @@
+from pennylane_calculquebec.exception import DeviceError
+
+class DeviceException(DeviceError):
+    """Exception for device-related errors (legacy alias)."""
+    pass
