+++ pennylane_calculquebec/processing/__init__.py
@@ -2,12 +2,4 @@
 
 from .monarq_postproc import PostProcessor
 from .monarq_preproc import PreProcessor
-
-from pennylane_calculquebec import PennylaneCQError as CalculQuebecError
-class ProcessingError(CalculQuebecError):
-    """Error related to processing."""
-    def __init__(self, message: str):
-        predefined = "Processing Error/"
-        full_message = f"{predefined}{message}"
-        super().__init__(full_message)
-        self.message = full_message
+from .processing_error import ProcessingError
