+++ pennylane_calculquebec/processing/processing_exception.py
@@ -1,0 +1,3 @@
+class ProcessingException(Exception):
+    """Exception raised for errors in the processing module."""
+    pass
