+++ pennylane_calculquebec/utility/utility_error.py
@@ -1,0 +1,8 @@
+class UtilityError(PennylaneCQError):
+    """Error related to utility."""
+    def __init__(self,):
+        predefined = "Utility Error/"
+        error_type=self.__class__.__name__
+        full_message = f"{predefined} of type {error_type}"
+        super().__init__(full_message)
+        self.message = full_message
